trusted-firmware-m: platform: nordic: rearrange veneers on builds without BL2
For builds without BL2, or when BL2 will be booting a single
combined S and NS image, we force placing the veneers section
at the end of the image (position 3), so as not to waste space
as a result of the nRF veneer section alignment requirements.
For regular builds with BL2 we keep the veneer placement as is
(positions 1 or 2 repending on whether PSA_API_TEST_NS is
defined).

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>